+++ app/src/abis/HalalPSM.ts
@@ -161,6 +161,39 @@ export const halalPsmAbi = [
   },
   {
     "type": "function",
+    "name": "cancelRedeemableWithPermit",
+    "inputs": [
+      {
+        "name": "hlcAmount",
+        "type": "uint256",
+        "internalType": "uint256"
+      },
+      {
+        "name": "deadline",
+        "type": "uint256",
+        "internalType": "uint256"
+      },
+      {
+        "name": "v",
+        "type": "uint8",
+        "internalType": "uint8"
+      },
+      {
+        "name": "r",
+        "type": "bytes32",
+        "internalType": "bytes32"
+      },
+      {
+        "name": "s",
+        "type": "bytes32",
+        "internalType": "bytes32"
+      }
+    ],
+    "outputs": [],
+    "stateMutability": "nonpayable"
+  },
+  {
+    "type": "function",
     "name": "cpiRate",
     "inputs": [],
     "outputs": [